Rename and update tests
